requirements.txt
pandas
numpy
tqdm
astroquery
astropy
scikit-learn
matplotlib
jupyter
torch
torchvision
grad-cam
Pillow
wandb
